backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
databases[asyncpg]==0.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
scipy==1.14.1
